fallback to error when android version is under 21

+++ lib/flutter_camera_ml_vision.dart
@@ -6,6 +6,7 @@ import 'dart:typed_data';
 import 'dart:ui';
 
 import 'package:camera/camera.dart';
+import 'package:device_info/device_info.dart';
 import 'package:firebase_ml_vision/firebase_ml_vision.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
@@ -14,11 +15,17 @@ import 'package:path_provider/path_provider.dart';
 part 'utils.dart';
 
 typedef HandleDetection<T> = Future<T> Function(FirebaseVisionImage image);
-typedef WidgetBuilder = Widget Function(BuildContext);
+typedef Widget ErrorWidgetBuilder(BuildContext context, CameraError error);
+
+enum CameraError {
+  unknown,
+  cantInitializeCamera,
+  androidVersionNotSupported,
+  noCameraAvailable,
+}
 
 enum _CameraState {
   loading,
-  noCamera,
   error,
   ready,
 }
@@ -27,7 +34,7 @@ class CameraMlVision<T> extends StatefulWidget {
   final HandleDetection<T> detector;
   final Function(T) onResult;
   final WidgetBuilder loadingBuilder;
-  final WidgetBuilder errorBuilder;
+  final ErrorWidgetBuilder errorBuilder;
 
   CameraMlVision({
     Key key,
@@ -47,6 +54,7 @@ class CameraMlVisionState<T> extends State<CameraMlVision<T>> {
   HandleDetection _detector;
   ImageRotation _rotation;
   _CameraState _cameraMlVisionState = _CameraState.loading;
+  CameraError _cameraError = CameraError.unknown;
   bool _alreadyCheckingImage = false;
   bool _isStreaming = false;
   bool _isDeactivate = false;
@@ -61,15 +69,19 @@ class CameraMlVisionState<T> extends State<CameraMlVision<T>> {
   }
 
   Future<void> stop() async {
-    if (_lastImage != null && File(_lastImage).existsSync()) {
-      await File(_lastImage).delete();
+    if(_cameraController != null) {
+      if (_lastImage != null && File(_lastImage).existsSync()) {
+        await File(_lastImage).delete();
+      }
+
+      Directory tempDir = await getTemporaryDirectory();
+      _lastImage = '${tempDir.path}/${DateTime
+          .now()
+          .millisecondsSinceEpoch}';
+      await _cameraController.takePicture(_lastImage);
+
+      await _stop(false);
     }
-
-    Directory tempDir = await getTemporaryDirectory();
-    _lastImage = '${tempDir.path}/${DateTime.now().millisecondsSinceEpoch}';
-    await _cameraController.takePicture(_lastImage);
-
-    await _stop(false);
   }
 
   Future<void> _stop(bool silently) async {
@@ -87,7 +99,9 @@ class CameraMlVisionState<T> extends State<CameraMlVision<T>> {
   }
 
   void start() {
-    _start(false);
+    if(_cameraController != null) {
+      _start(false);
+    }
   }
 
   void _start(bool silently) {
@@ -102,9 +116,24 @@ class CameraMlVisionState<T> extends State<CameraMlVision<T>> {
   }
 
   Future<void> _initialize() async {
+    if (Platform.isAndroid) {
+      final deviceInfo = DeviceInfoPlugin();
+      final androidInfo = await deviceInfo.androidInfo;
+      if (androidInfo.version.sdkInt < 21) {
+        debugPrint('Camera plugin doesn\'t support android under version 21');
+        setState(() {
+          _cameraMlVisionState = _CameraState.error;
+          _cameraError = CameraError.androidVersionNotSupported;
+        });
+        return;
+      }
+    }
+
     CameraDescription description = await _getCamera(CameraLensDirection.back);
     if (description == null) {
-      _cameraMlVisionState = _CameraState.noCamera;
+      _cameraMlVisionState = _CameraState.error;
+      _cameraError = CameraError.noCameraAvailable;
+
       return;
     }
     _cameraController = CameraController(description,
@@ -118,11 +147,13 @@ class CameraMlVisionState<T> extends State<CameraMlVision<T>> {
     } catch (ex, stack) {
       setState(() {
         _cameraMlVisionState = _CameraState.error;
+        _cameraError = CameraError.cantInitializeCamera;
       });
       debugPrint('Can\'t initialize camera');
       debugPrint('$ex, $stack');
       return;
     }
+
     setState(() {
       _cameraMlVisionState = _CameraState.ready;
     });
@@ -135,12 +166,14 @@ class CameraMlVisionState<T> extends State<CameraMlVision<T>> {
 
   @override
   void deactivate() {
-    if (_isDeactivate) {
-      _isDeactivate = false;
-      _start(true);
-    } else {
-      _isDeactivate = true;
-      _stop(true);
+    if(_cameraController != null) {
+      if (_isDeactivate) {
+        _isDeactivate = false;
+        _start(true);
+      } else {
+        _isDeactivate = true;
+        _stop(true);
+      }
     }
     super.deactivate();
   }
@@ -150,7 +183,9 @@ class CameraMlVisionState<T> extends State<CameraMlVision<T>> {
     if (_lastImage != null && File(_lastImage).existsSync()) {
       File(_lastImage).delete();
     }
-    _cameraController.dispose();
+    if(_cameraController != null) {
+      _cameraController.dispose();
+    }
     super.dispose();
   }
 
@@ -161,11 +196,10 @@ class CameraMlVisionState<T> extends State<CameraMlVision<T>> {
           ? Center(child: CircularProgressIndicator())
           : widget.loadingBuilder(context);
     }
-    if (_cameraMlVisionState == _CameraState.noCamera ||
-        _cameraMlVisionState == _CameraState.error) {
+    if (_cameraMlVisionState == _CameraState.error) {
       return widget.errorBuilder == null
-          ? Center(child: Text('$_cameraMlVisionState'))
-          : widget.errorBuilder(context);
+          ? Center(child: Text('$_cameraMlVisionState $_cameraError'))
+          : widget.errorBuilder(context, _cameraError);
     }
     return FittedBox(
       alignment: Alignment.center,
